Update Wheelchair_Operation.py
This was the main script to be run for the wheelchairs operation.  time constraints, working with new languages, hardware restrictions, and the schedule of a college semester resulted in some unpolished code. That being said the system worked for testing purposes and served as a successful proof of concept. The basic operation of the program is as follows. Upon starting the wheelchair's Raspberry Pi the script will automatically be entered and will generate a tkinter interface. This interface is extremely simple and only consists of a few buttons. These buttons represent the mode you wish to have the wheelchair move to. 
The modes are: Mobile, Horizontal Rails Extended, Paws Extended, Vertical Rails Extended, and Seat Rise.
These modes must stay in that order due to mechanical constraints. Based on the previous mode the script will determine what actions need to be taken to reach the desired mode. Due to hardware restrictions no sensors were able to be added to detect current mode, instead the program forces the wheelchair to reset to Mobile during the startup sequence. 
The actuation is all based on timing, again we had limited sensors so the simplest method to make a working prototype involved timing out each of the actuators to allow for complete extension (slightly longer to ensure full retraction).